+++ src/Equinor.Procosys.Preservation.MainApi/Tag/ProcosysTagOverview.cs
@@ -1,0 +1,12 @@
+﻿using System.Diagnostics;
+
+namespace Equinor.Procosys.Preservation.MainApi.Tag
+{
+    [DebuggerDisplay("{TagNo}")]
+    public class ProcosysTagOverview
+    {
+        public long Id { get; set; }
+        public string TagNo { get; set; }
+        public string Description { get; set; }
+    }
+}
